Match tzfile(5)'s description of format for table vs. rule.
tzfile(5) says the rule, when present, is for times after the last
transition in the file.  (I use it also for exactly that time, to avoid
extra discontinuity in handling the exact transition time.)  Honor this
rule rather than assuming how zic was used to compile the zoneinfo; it
will be important when time zone rules change after 2038.