Fix handling of UTC-offset abbreviations in rules.
The timezone database introduced (after a good bit of planning) the use
of UTC offsets as timezone abbreviations, to replace the practice of
making up abbreviations when local practice did not exist.  This was
done in git commit 8ee11a301cf173afb0c76e0315b9f9ec8ebb9d95 on
2016-04-15.

This change is needed to handle the <+06> form of rules.  The <> should
be stripped from the abbreviation in the rule case, since it is not
present in the non-rule case.